move files to eth directory
