+++ protocol/mcr/cli/client/src/cli/eth/get_accepted_commitment_at_height/mod.rs_OLD
@@ -58,7 +58,7 @@ impl CheckPostconfirmation {
 
         let client = config.build().await?;
 
-        let commitment = client.get_commitment_at_height(
+        let commitment = client.get_accepted_commitment_at_height(
             self.args.height
         ).await?;
