+++ protocol/mcr/cli/client/src/cli/eth/get_stake/mod.rs
