+++ protocol/mcr/cli/client/src/cli/eth/mod.rs
@@ -1,13 +1,15 @@
-pub mod get_commitment_at_height;
+pub mod get_commitment;
+pub mod get_accepted_commitment_at_height;
 pub mod get_max_tolerable_block_height;
 pub mod get_posted_commitment_at_height;
 pub mod post_admin_commitment;
 pub mod post_commitment;
 pub mod post_commitment_batch;
 pub mod stake;
+pub mod get_stake;
 pub mod stream_commitments;
 pub mod unstake;
-
+pub mod grant_trusted_attester;
 use clap::Subcommand;
 
 #[derive(Subcommand)]
@@ -22,30 +24,39 @@ pub enum Eth {
 	PostCommitmentBatch(post_commitment_batch::PostCommitmentBatch),
 	/// Stream commitments
 	StreamCommitments(stream_commitments::StreamCommitments),
-	/// Get commitment at a specific height
-	GetCommitmentAtHeight(get_commitment_at_height::GetCommitmentAtHeight),
+	/// Get a commitment for a given height and attester
+	GetCommitment(get_commitment::GetCommitment),
+	/// Get accepted commitment at a specific height
+	GetAcceptedCommitmentAtHeight(get_accepted_commitment_at_height::GetAcceptedCommitmentAtHeight),
 	/// Get posted commitment at a specific height
 	GetPostedCommitmentAtHeight(get_posted_commitment_at_height::GetPostedCommitmentAtHeight),
 	/// Get max tolerable superBlock height
 	GetMaxTolerableBlockHeight(get_max_tolerable_block_height::GetMaxTolerableBlockHeight),
 	/// Stake tokens for the MCR domain
 	Stake(stake::Stake),
+	/// Get the current epoch stake for an attester
+	GetStake(get_stake::GetStake),
 	/// Unstake tokens from the MCR domain
 	Unstake(unstake::Unstake),
+	/// Grant TRUSTED_ATTESTER role to an attester
+	GrantTrustedAttester(grant_trusted_attester::GrantTrustedAttester),
 }
 
 impl Eth {
 	pub async fn execute(&self) -> Result<(), anyhow::Error> {
 		match self {
+			Eth::GetCommitment(cmd) => cmd.execute().await?,
 			Eth::PostAdminCommitment(cmd) => cmd.execute().await?,
 			Eth::PostCommitment(cmd) => cmd.execute().await?,
 			Eth::PostCommitmentBatch(cmd) => cmd.execute().await?,
 			Eth::StreamCommitments(cmd) => cmd.execute().await?,
-			Eth::GetCommitmentAtHeight(cmd) => cmd.execute().await?,
+			Eth::GetAcceptedCommitmentAtHeight(cmd) => cmd.execute().await?,
 			Eth::GetPostedCommitmentAtHeight(cmd) => cmd.execute().await?,
 			Eth::GetMaxTolerableBlockHeight(cmd) => cmd.execute().await?,
 			Eth::Stake(cmd) => cmd.execute().await?,
+			Eth::GetStake(cmd) => cmd.execute().await?,
 			Eth::Unstake(cmd) => cmd.execute().await?,
+			Eth::GrantTrustedAttester(cmd) => cmd.execute().await?,
 		}
 		Ok(())
 	}
