+++ protocol/mcr/clients/eth/src/client/mod.rs
@@ -187,7 +187,7 @@ where
 		Ok(Box::pin(stream) as CommitmentStream)
 	}
 
-	async fn get_commitment_at_height(
+	async fn get_accepted_commitment_at_height(
 		&self,
 		height: u64,
 	) -> Result<Option<BlockCommitment>, McrClientError> {
