+++ protocol/mcr/clients/mock/src/lib.rs
@@ -133,7 +133,7 @@ impl McrClientOperations for Client {
 		Ok(Box::pin(ReceiverStream::new(receiver)))
 	}
 
-	async fn get_commitment_at_height(
+	async fn get_accepted_commitment_at_height(
 		&self,
 		height: u64,
 	) -> Result<Option<BlockCommitment>, McrClientError> {
