+++ protocol/pcp/clients/eth/src/client/mod.rs
@@ -182,7 +182,7 @@ where
 		Ok(Box::pin(stream) as CommitmentStream)
 	}
 
-	async fn get_commitment_at_height(
+	async fn get_accepted_commitment_at_height(
 		&self,
 		height: u64,
 	) -> Result<Option<SuperBlockCommitment>, PcpClientError> {
